@@ -1,3 +1,3 @@
 module github.com/spaceship/date
 
-go 1.12
+go 1.18
