@@ -1,7 +1,10 @@
 package date
 
 import (
+	"bytes"
+	"database/sql"
 	"database/sql/driver"
+	"encoding/json"
 	"fmt"
 	"time"
 )
@@ -209,7 +212,7 @@ func (d *Date) UnmarshalJSON(p []byte) error {
 
 // Scan implements the sql.Scanner interface, allowing the sql package to read
 // sql dates into Date.
-func (d *Date) Scan(src interface{}) error {
+func (d *Date) Scan(src any) error {
 	t, ok := src.(time.Time)
 	if !ok {
 		return fmt.Errorf("can not scan as Date: %T", src)
@@ -240,4 +243,70 @@ func Min(d1, d2 Date) Date {
 		return d1
 	}
 	return d2
+}
+
+// NullDate represents a nullable date with compatibility for
+// scanning null values from the database.
+type NullDate struct {
+	Date  Date
+	Valid bool
+}
+
+func NewNullDate(d Date) NullDate {
+	return NullDate{
+		Date:  d,
+		Valid: true,
+	}
+}
+
+// NullDate should properly implement Scan and Value
+type NullDateSqlField interface {
+	sql.Scanner
+	driver.Valuer
+}
+
+var _ NullDateSqlField = NullDate{}
+
+// NullDate should properly implement UnmarshalJSON and MarshalJSON
+type NullDateJsonField interface {
+	json.Unmarshaler
+	json.Marshaler
+}
+
+var _ NullDateJsonField = NullDate{}
+
+// Scan implements the sql.Scanner interface for database deserialization.
+func (d NullDate) Scan(value any) error {
+	if value == nil {
+		d.Valid = false
+		return nil
+	}
+	d.Valid = true
+	return d.Date.Scan(value)
+}
+
+// Value implements the driver.Valuer interface for database serialization.
+func (d NullDate) Value() (driver.Value, error) {
+	if !d.Valid {
+		return nil, nil
+	}
+	return d.Date.Value()
+}
+
+// UnmarshalJSON implements the json.Unmarshaler interface.
+func (d NullDate) UnmarshalJSON(dateBytes []byte) error {
+	if bytes.Equal(dateBytes, []byte("null")) {
+		d.Valid = false
+		return nil
+	}
+	d.Valid = true
+	return d.Date.UnmarshalJSON(dateBytes)
+}
+
+// MarshalJSON implements the json.Marshaler interface.
+func (d NullDate) MarshalJSON() ([]byte, error) {
+	if !d.Valid {
+		return []byte("null"), nil
+	}
+	return d.Date.MarshalJSON()
 }
